% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saluda.R
\name{saluda}
\alias{saluda}
\title{saluda}
\usage{
saluda(name)
}
\arguments{
\item{name}{tu nombre}
}
\value{
una cadena de texto con el saludo personalizado
}
\description{
Dice hola a la persona pasada por parámetro
}
\examples{
saluda("Luz")
saluda("Pepe")
}

