% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.dos.numeros.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Calcula la suma de dos numeros}
\usage{
suma.dos.numeros(x, y)
}
\arguments{
\item{x}{un operando}

\item{y}{otro operando}
}
\value{
el resultado de sumar x e y
}
\description{
Calcula la suma de dos numeros
}
\examples{
suma.dos.numeros(2, 3)
suma.dos.numeros(4, 4)
}

